Final Push: Ready for Review
Ready For review. On the home page, affected both row and column height and width respectively. Also affected to text slightly. On the other page, divided the page into two columns, added a line in the center, affected the number of widows and orphans that will appear, and made the title of it span across both columns.